Integration with GUL
This integration will work with any plugins/themes that have data in the Update URI header, only requirement is AspireUpdate be active.

There is an update to GUL coming but the code is all in here, it's just incorrectly versioned so don't run composer update until GUL v2.2.0 is released.