Admin_Settings: Display notices instead of alerts when managing the log.
Previously, when clearing the log or encountering an error viewing the log,
a message would appear in an alert. This change generates admin notices instead.

The notice when clearing the log, and the notice when viewing a log, each have their own ID. Repeated clicks of a button will simply replace the existing respective notice to avoid notice spam.